+++ js/functions.js
@@ -1,21 +1,15 @@
-/*Функция для проверки длины строки. Она принимает строку, которую нужно проверить, и максимальную длину и возвращает true, если строка меньше или равна указанной длине, и false, если строка длиннее.
-
-Для решения этой задачи вам потребуется объявить функцию с двумя параметрами: строкой и максимальной длиной. В теле функции используйте оператор сравнения меньше или равно (<=), чтобы сравнить длину полученной строки (свойство length) с максимальной длиной. Функция должна вернуть результат этого сравнения.*/
-
+/*Функция для проверки длины строки. Она принимает строку, которую нужно проверить, и максимальную длину и возвращает true, если строка меньше или равна указанной длине, и false, если строка длиннее.*/
 
 const checkStringLength = (string, maxLength) => string.length <= maxLength;
 
 // Cтрока короче 20 символов
-checkStringLength('проверяемая строка', 20);
-// true
+checkStringLength('проверяемая строка', 20); // true
 
 // Длина строки ровно 18 символов
-checkStringLength('проверяемая строка', 18);
-// true
+checkStringLength('проверяемая строка', 18); // true
 
 // Строка длиннее 10 символов
-checkStringLength('проверяемая строка', 10);
-// false
+checkStringLength('проверяемая строка', 10); // false
 
 const isPalindrome = (string) => {
   const normalizedString = string.toLowerCase().replaceAll(' ', '');
@@ -29,11 +23,10 @@ const isPalindrome = (string) => {
 };
 
 // Строка является палиндромом
-isPalindrome('топот');
-// true
+isPalindrome('топот'); // true
+
 // Несмотря на разный регистр, тоже палиндром
-isPalindrome('ДовОд');
-// true
+isPalindrome('ДовОд'); // true
+
 // Это не палиндром
-isPalindrome('Кекс');
-// false
+isPalindrome('Кекс'); // false
